Moved a couple of files around into standard format
